[ref #12] Refactor `Electrolyzer`
This commit extracts all electrolyzer operations from electrochemical
processes. The resulting classes are `ElectrolyzerStack` and
`ElecrolyzerCell`.
