+++ examples/example_01_electrolyzer/run_electrolyzer.py
@@ -4,7 +4,8 @@ This example runs the Electrolyzer on its own and generates polarization curves.
 import numpy as np
 import matplotlib.pyplot as plt
 
-from electrolyzer.electrolyzer import Electrolyzer, electrolyzer_model
+from electrolyzer.electrolyzer_cell import electrolyzer_model
+from electrolyzer.electrolyzer_stack import ElectrolyzerStack
 
 
 n_cells = 100  # number of cells in stack
@@ -12,27 +13,15 @@ cell_area = 1000  # cell area, cm^2
 temperature = 60  # temperature
 max_current = 2000
 
-elec = Electrolyzer(n_cells, cell_area, temperature, max_current, dt=1)
+elec = ElectrolyzerStack(n_cells, cell_area, temperature, max_current, dt=1)
 
 cur = np.linspace(0, 2500, 100)
-fit_error = np.zeros_like(cur)
-p_actual = np.zeros_like(cur)
-p_fit = np.zeros_like(cur)
-voltage = np.zeros_like(cur)
-
-for i in range(len(cur)):
-    fit_error[i] = elec.calc_stack_power(cur[i]) - elec.calc_stack_power(
-        electrolyzer_model(
-            (elec.calc_stack_power(cur[i]), temperature), *elec.fit_params
-        )
-    )
-    p_actual[i] = elec.calc_stack_power(cur[i])
-    p_fit[i] = elec.calc_stack_power(
-        electrolyzer_model(
-            (elec.calc_stack_power(cur[i]), temperature), *elec.fit_params
-        )
-    )
-    voltage[i] = elec.calc_cell_voltage(cur[i])
+p_fit = elec.calc_stack_power(
+    electrolyzer_model((elec.calc_stack_power(cur), temperature), *elec.fit_params)
+)
+p_actual = elec.calc_stack_power(cur)
+voltage = elec.cell.calc_cell_voltage(cur, temperature)
+fit_error = p_actual - p_fit
 
 fig, (ax1, ax2) = plt.subplots(2, 1, sharex=True)
